UI: made the sidebar nav items cleaner and more consistent
This now makes it so all nav items use the `nav_item` or new `nav_list_item` helpers.
