+++ app/webpack/stylesheets/_sidebar.scss
@@ -18,16 +18,13 @@
   }
 
   .sidebar-heading {
-    @extend .px-3;
     @extend .mb-2;
     @extend .text-muted;
     font-size: .85rem;
     text-transform: uppercase;
-    margin:0 !important;
 
     .btn-sm {
       font-size: 0.75em;
-      margin-top: -5px;
     }
   }
